include: sys: Add back some XNU headers with fixes and edits
Stripped away private KPI function declarations since we can't use them.

Edits:
	- sys/proc.h: Removed function declarations that are not supported by
  	  the public KPI and fixed others by adding the 'extern' prefix,
	  added in the flags and other definitions from the private KPI.
	  This should now overwrite the buggy one in the MacKernelSDK.
	- sys/proc_internal.h: Cleanup, fixes and edits.
	- sys/vnode_internal.h: Minor cosmetic edits.

Headers added:
	- mach/coalition.h
	- sys/bsdtask_info.h (with edits)
	- sys/event.h
	- sys/file_internal.h (with edits)
	- sys/guarded.h
	- sys/reason.h
	- sys/tty.h (with edits)
	- sys/ttydefaults.h

It's nice to have these for r&d as well as the various flags, macros and
definitions they provide, not to mention the structures even though the
goal is to depend on them as little as possible in the end.