kext: lib/cpu: Tidy things up and add function set_microcode_version
set_microcode_version() will cause a kernel panic at the moment if used.
I'm trying to get to the bottom of why the cpuid_microcode_version field in
struct i386_cpu_info has to be set to 186 in the AMD kernel in order to boot.

/proc/cpuinfo can still use get_microcode_version to get the correct value.
However it doesn't solve the underlying issue of why cpuid_microcode_version
in struct i386_cpu_info needs to be spoofed in the first place. I want to try
and figure out how it can be assigned the correct value without causing boot
issues.

Other changes in this commit:

- Moved cexp[3] to the cpu library.
- Changed the type for get_microcode_version to uint32_t.